pyproject.toml
[project]
name = "garmin-lt-public"
version = "2025.11.0"
description = "Self-hosted permanent Garmin LiveTrack web application"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "flask>=3.0.0",
    "google-auth>=2.25.0",
    "google-auth-oauthlib>=1.2.0",
    "google-auth-httplib2>=0.2.0",
    "google-api-python-client>=2.110.0",
    "beautifulsoup4>=4.12.0",
    "python-dotenv>=1.0.0",
    "gunicorn>=21.2.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
]
